mlady.dm
Very light brute damage compliment, the first useful virus food trait.
By proxy, it's actually hard to manifest! It'll take multiple tries.
I've tested it, and it's not very dangerous. To compensate for it being a minor trait to add to far more lethal viruses, it's only less than the strength of something like flesh-eating, and takes longer to manifest.